Update scrypt.php
I thought it would be better to make the `check` method more portable by just getting the key length from the hash as opposed to always using a hard-coded `$_keyLength` property. As it is now, if someone decides to change that property, any hashes created prior to that will not verify correctly.

Note: The `$isShadowed` variable is `false` on my server. I'm not sure if this would make a difference in calculating the string length in a way that would cause it to not get the right length to put in the hash function. Or maybe it's possible just to use `strlen` instead of `self::strlen`, in the lines I changed, but I am unable to test it.